refactor(cli-engine): split the monoliths into focused modules; ProductManifest becomes CommandFamily
core.ts and execution.ts carried the whole engine as banner-comment
sections. The definition surface now lives in src/definition/ (args,
commands, config-section, context, envelopes, events, presentation,
runtime, streams, command-family) and the runner in src/execution/
(command-tree, shared-flags, stricli-adapter, needs, prompts,
command-context, events, render, settlement, run, harness), every file
well under 500 lines, with the entry modules under src/exports/.
Banner comments are gone; module boundaries carry the structure.

Operator ruling on vocabulary: the unit of contribution a package
exports is a CommandFamily. ProductManifest is renamed to CommandFamily,
createCli's products key to commandFamilies, and the internal mount
entry/node/tree names adopt command-tree vocabulary (CommandTreeEntry,
CommandTreeNode, buildCommandTree); 'mount' survives only as the verb.

No behavior change; both test suites pass unchanged in count.
