+++ packages/cli-engine/tsdown.config.ts
@@ -2,13 +2,12 @@ import { defineConfig } from "tsdown";
 
 export default defineConfig({
   entry: {
-    index: "src/index.ts",
-    protocol: "src/protocol.ts",
+    index: "src/exports/index.ts",
+    protocol: "src/exports/protocol.ts",
   },
   format: ["esm"],
   dts: true,
   clean: true,
   fixedExtension: false,
   outDir: "dist",
-  deps: { neverBundle: ["@prisma/cli-engine/protocol"] },
 });
